+++ R/datasummary_correlation.R
@@ -195,9 +195,14 @@ datasummary_correlation <- function(data,
     title = title,
     ...)
 
-  if (!is.null(settings_get("output_file"))) {
+
+  # invisible return
+  if (!is.null(settings_get("output_file")) ||
+      output == "jupyter" ||
+      (output == "default" && settings_equal("output_default", "jupyter"))) {
     settings_rm()
     return(invisible(out))
+  # visible return
   } else {
     settings_rm()
     return(out)